requirements.txt
# API und Daten
alpaca-trade-api>=3.0.0
pandas>=1.3.0
numpy>=1.21.0

# Reinforcement Learning
gymnasium>=0.28.1
torch>=1.10.0

# Visualisierung
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0

# Utilities
python-dotenv>=0.19.0